requirements.txt
appdirs==1.4.3
CacheControl==0.12.6
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
decorator==4.4.2
Deprecated==1.2.10
distlib==0.3.0
distro==1.4.0
ffmpeg==1.4
ffmpeg-python==0.2.0
future==0.18.2
gitdb==4.0.5
html5lib==1.0.1
idna==2.10
imageio==2.9.0
imageio-ffmpeg==0.4.2
ipaddr==2.2.0
lockfile==0.12.2
moviepy==1.0.3
msgpack==0.6.2
numpy==1.19.0
oauthlib==3.1.0
packaging==20.3
pep517==0.8.2
Pillow==7.2.0
proglog==0.1.9
progress==1.5
PyGithub==1.51
PyJWT==1.7.1
pyparsing==2.4.6
PySocks==1.7.1
pytoml==0.1.21
requests==2.24.0
requests-oauthlib==1.3.0
retrying==1.3.3
six==1.15.0
smmap==3.0.4
tqdm==4.48.0
tweepy==3.9.0
urllib3==1.25.9
webencodings==0.5.1
wrapt==1.12.1
